update and bugfixes

+++ DatabaseCodeGenerator/Schema/Extensions.cs
@@ -10,8 +10,8 @@
 
         public static void MinMaxFloat(this XmlClassMember member, out float minvalue, out float maxvalue)
         {
-            if (string.IsNullOrEmpty(member.minvalue) || !float.TryParse(member.minvalue, out minvalue)) minvalue = int.MinValue;
-            if (string.IsNullOrEmpty(member.maxvalue) || !float.TryParse(member.maxvalue, out maxvalue)) maxvalue = int.MaxValue;
+            if (string.IsNullOrEmpty(member.minvalue) || !float.TryParse(member.minvalue, out minvalue)) minvalue = float.MinValue;
+            if (string.IsNullOrEmpty(member.maxvalue) || !float.TryParse(member.maxvalue, out maxvalue)) maxvalue = float.MaxValue;
         }
     }
 }
